Refactored to move Storer interface to Types
