+++ 06_puppy/srini/Types.go
@@ -12,3 +12,11 @@ type Puppy struct {
 	colour string
 	value  float64
 }
+
+//Storer provides CRUD methods
+type Storer interface {
+	CreatePuppy(*Puppy) int
+	ReadPuppy(int) (*Puppy, error)
+	UpdatePuppy(int, *Puppy) error
+	DeletePuppy(int) error
+}
